[video] CVideoInfoTag: Remove CVideoDataBase dependency. Improves performance of ListItem.IsVideoExtras GUI label implementation drastically, because no more db lookup needed.
