+++ xbmc/video/CMakeLists.txt
@@ -36,6 +36,7 @@ set(HEADERS Bookmark.h
             VideoLibraryQueue.h
             VideoThumbLoader.h
             VideoUtils.h
+            VideoVersionTypes.h
             ViewModeSettings.h)
 
 core_add_library(video)
